feat: same for publish
