+++ crates/sampo-core/src/lib.rs
@@ -2,6 +2,7 @@ pub mod changeset;
 pub mod config;
 pub mod enrichment;
 pub mod errors;
+pub mod publish;
 pub mod release;
 pub mod types;
 pub mod workspace;
@@ -17,6 +18,10 @@ pub use enrichment::{
     enrich_changeset_message, get_commit_hash_for_path,
 };
 pub use errors::SampoError;
+pub use publish::{
+    is_publishable_to_crates_io, run_publish, tag_published_crate, topo_order,
+    version_exists_on_crates_io,
+};
 pub use release::{
     build_dependency_updates, bump_version, create_dependency_update_entry,
     create_fixed_dependency_policy_entry, detect_all_dependency_explanations,
