+++ crates/sampo-core/src/publish.rs
@@ -1,0 +1,348 @@
+use crate::types::CrateInfo;
+use crate::discover_workspace;
+use std::collections::{BTreeMap, BTreeSet, VecDeque};
+use std::fs;
+use std::io;
+use std::path::Path;
+use std::process::Command;
+use std::time::Duration;
+
+/// Main publish function that can be called from CLI or other interfaces
+pub fn run_publish(root: &std::path::Path, dry_run: bool, cargo_args: &[String]) -> io::Result<()> {
+    let ws = discover_workspace(root).map_err(io::Error::other)?;
+
+    // Determine which crates are publishable to crates.io
+    let mut name_to_crate: BTreeMap<String, &CrateInfo> = BTreeMap::new();
+    let mut publishable: BTreeSet<String> = BTreeSet::new();
+    for c in &ws.members {
+        let manifest = c.path.join("Cargo.toml");
+        if is_publishable_to_crates_io(&manifest)? {
+            publishable.insert(c.name.clone());
+            name_to_crate.insert(c.name.clone(), c);
+        }
+    }
+
+    if publishable.is_empty() {
+        println!("No publishable crates for crates.io were found in the workspace.");
+        return Ok(());
+    }
+
+    // Validate internal deps do not include non-publishable crates
+    let mut errors: Vec<String> = Vec::new();
+    for name in &publishable {
+        let c = name_to_crate.get(name).unwrap();
+        for dep in &c.internal_deps {
+            if !publishable.contains(dep) {
+                errors.push(format!(
+                    "crate '{}' depends on internal crate '{}' which is not publishable",
+                    name, dep
+                ));
+            }
+        }
+    }
+    if !errors.is_empty() {
+        for e in errors {
+            eprintln!("{e}");
+        }
+        return Err(io::Error::other(
+            "cannot publish due to non-publishable internal dependencies",
+        ));
+    }
+
+    // Compute publish order (topological: deps first) for all publishable crates.
+    let order = topo_order(&name_to_crate, &publishable).map_err(io::Error::other)?;
+
+    println!("Publish plan (crates.io):");
+    for name in &order {
+        println!("  - {name}");
+    }
+
+    // Execute cargo publish in order
+    for name in &order {
+        let c = name_to_crate.get(name).unwrap();
+        let manifest = c.path.join("Cargo.toml");
+        // Skip if the exact version already exists on crates.io
+        match version_exists_on_crates_io(&c.name, &c.version) {
+            Ok(true) => {
+                println!(
+                    "Skipping {}@{} (already exists on crates.io)",
+                    c.name, c.version
+                );
+                continue;
+            }
+            Ok(false) => {}
+            Err(e) => {
+                eprintln!(
+                    "Warning: could not check crates.io for {}@{}: {}. Attempting publish…",
+                    c.name, c.version, e
+                );
+            }
+        }
+
+        let mut cmd = Command::new("cargo");
+        cmd.arg("publish").arg("--manifest-path").arg(&manifest);
+        if dry_run {
+            cmd.arg("--dry-run");
+        }
+        if !cargo_args.is_empty() {
+            cmd.args(cargo_args);
+        }
+
+        println!(
+            "Running: {}",
+            format_command_display(cmd.get_program(), cmd.get_args())
+        );
+
+        let status = cmd.status()?;
+        if !status.success() {
+            return Err(io::Error::other(format!(
+                "cargo publish failed for crate '{}' with status {}",
+                name, status
+            )));
+        }
+
+        // Create an annotated git tag after successful publish (not in dry-run)
+        if !dry_run
+            && let Err(e) = tag_published_crate(&ws.root, &c.name, &c.version)
+        {
+            eprintln!(
+                "Warning: failed to create tag for {}@{}: {}",
+                c.name, c.version, e
+            );
+        }
+    }
+
+    if dry_run {
+        println!("Dry-run complete.");
+    } else {
+        println!("Publish complete.");
+    }
+
+    Ok(())
+}
+
+/// Check if a crate is publishable to crates.io based on its manifest
+pub fn is_publishable_to_crates_io(manifest_path: &Path) -> io::Result<bool> {
+    let text = fs::read_to_string(manifest_path)?;
+    let value: toml::Value = text
+        .parse()
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, format!("{e}")))?;
+
+    let pkg = match value.get("package").and_then(|v| v.as_table()) {
+        Some(p) => p,
+        None => return Ok(false),
+    };
+
+    // If publish = false => skip
+    if let Some(val) = pkg.get("publish") {
+        match val {
+            toml::Value::Boolean(false) => return Ok(false),
+            toml::Value::Array(arr) => {
+                // Only publish if the array contains "crates-io"
+                // (Cargo uses this to whitelist registries.)
+                let allowed: Vec<String> = arr
+                    .iter()
+                    .filter_map(|v| v.as_str().map(|s| s.to_string()))
+                    .collect();
+                return Ok(allowed.iter().any(|s| s == "crates-io"));
+            }
+            _ => {}
+        }
+    }
+
+    // Default case: publishable
+    Ok(true)
+}
+
+/// Create a git tag for a published crate
+pub fn tag_published_crate(repo_root: &Path, crate_name: &str, version: &str) -> io::Result<()> {
+    if !repo_root.join(".git").exists() {
+        // Not a git repo, skip
+        return Ok(());
+    }
+    let tag = format!("{}-v{}", crate_name, version);
+    // If tag already exists, do not recreate
+    let out = Command::new("git")
+        .arg("-C")
+        .arg(repo_root)
+        .arg("tag")
+        .arg("--list")
+        .arg(&tag)
+        .output()?;
+    if out.status.success() {
+        let s = String::from_utf8_lossy(&out.stdout);
+        if s.lines().any(|l| l.trim() == tag) {
+            return Ok(());
+        }
+    }
+
+    let msg = format!("Release {} {}", crate_name, version);
+    let status = Command::new("git")
+        .arg("-C")
+        .arg(repo_root)
+        .arg("tag")
+        .arg("-a")
+        .arg(&tag)
+        .arg("-m")
+        .arg(&msg)
+        .status()?;
+    if status.success() {
+        Ok(())
+    } else {
+        Err(io::Error::other(format!(
+            "git tag failed with status {}",
+            status
+        )))
+    }
+}
+
+/// Check if a version exists on crates.io
+pub fn version_exists_on_crates_io(crate_name: &str, version: &str) -> io::Result<bool> {
+    // Query crates.io: https://crates.io/api/v1/crates/<name>/<version>
+    let url = format!("https://crates.io/api/v1/crates/{}/{}", crate_name, version);
+
+    let client = reqwest::blocking::Client::builder()
+        .timeout(Duration::from_secs(10))
+        .user_agent("sampo/0.4.0")
+        .build()
+        .map_err(|e| io::Error::other(format!("failed to build HTTP client: {}", e)))?;
+
+    let res = client
+        .get(&url)
+        .send()
+        .map_err(|e| io::Error::other(format!("HTTP request failed: {}", e)))?;
+
+    let status = res.status();
+    if status == reqwest::StatusCode::OK {
+        Ok(true)
+    } else if status == reqwest::StatusCode::NOT_FOUND {
+        Ok(false)
+    } else {
+        // Include a short, normalized snippet of the response body for diagnostics
+        let body = res.text().unwrap_or_default();
+        let snippet: String = body.trim().chars().take(500).collect();
+        let snippet = snippet.split_whitespace().collect::<Vec<_>>().join(" ");
+
+        let body_part = if snippet.is_empty() {
+            String::new()
+        } else {
+            format!(" body=\"{}\"", snippet)
+        };
+
+        Err(io::Error::other(format!(
+            "Crates.io {} response: {}",
+            status, body_part
+        )))
+    }
+}
+
+/// Compute topological order for publishing crates (dependencies first)
+pub fn topo_order(
+    name_to_crate: &BTreeMap<String, &CrateInfo>,
+    include: &BTreeSet<String>,
+) -> Result<Vec<String>, String> {
+    // Build graph: edge dep -> crate
+    let mut indegree: BTreeMap<&str, usize> = BTreeMap::new();
+    let mut forward: BTreeMap<&str, Vec<&str>> = BTreeMap::new();
+
+    for name in include {
+        indegree.insert(name.as_str(), 0);
+        forward.entry(name.as_str()).or_default();
+    }
+
+    for name in include {
+        let c = name_to_crate
+            .get(name)
+            .ok_or_else(|| format!("missing crate info for '{}'", name))?;
+        for dep in &c.internal_deps {
+            if include.contains(dep) {
+                // dep -> name
+                let entry = forward.entry(dep.as_str()).or_default();
+                entry.push(name.as_str());
+                *indegree.get_mut(name.as_str()).unwrap() += 1;
+            }
+        }
+    }
+
+    let mut q: VecDeque<&str> = indegree
+        .iter()
+        .filter_map(|(k, &d)| if d == 0 { Some(*k) } else { None })
+        .collect();
+    let mut out: Vec<String> = Vec::new();
+
+    while let Some(n) = q.pop_front() {
+        out.push(n.to_string());
+        if let Some(children) = forward.get(n) {
+            for &m in children {
+                if let Some(d) = indegree.get_mut(m) {
+                    *d -= 1;
+                    if *d == 0 {
+                        q.push_back(m);
+                    }
+                }
+            }
+        }
+    }
+
+    if out.len() != include.len() {
+        return Err("dependency cycle detected among publishable crates".into());
+    }
+    Ok(out)
+}
+
+fn format_command_display(program: &std::ffi::OsStr, args: std::process::CommandArgs) -> String {
+    let prog = program.to_string_lossy();
+    let mut s = String::new();
+    s.push_str(&prog);
+    for a in args {
+        s.push(' ');
+        s.push_str(&a.to_string_lossy());
+    }
+    s
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::path::PathBuf;
+
+    #[test]
+    fn topo_orders_deps_first() {
+        // Build a small fake graph using CrateInfo structures
+        let a = CrateInfo {
+            name: "a".into(),
+            version: "0.1.0".into(),
+            path: PathBuf::from("/tmp/a"),
+            internal_deps: BTreeSet::new(),
+        };
+        let mut deps_b = BTreeSet::new();
+        deps_b.insert("a".into());
+        let b = CrateInfo {
+            name: "b".into(),
+            version: "0.1.0".into(),
+            path: PathBuf::from("/tmp/b"),
+            internal_deps: deps_b,
+        };
+        let mut deps_c = BTreeSet::new();
+        deps_c.insert("b".into());
+        let c = CrateInfo {
+            name: "c".into(),
+            version: "0.1.0".into(),
+            path: PathBuf::from("/tmp/c"),
+            internal_deps: deps_c,
+        };
+
+        let mut map: BTreeMap<String, &CrateInfo> = BTreeMap::new();
+        map.insert("a".into(), &a);
+        map.insert("b".into(), &b);
+        map.insert("c".into(), &c);
+
+        let mut include = BTreeSet::new();
+        include.insert("a".into());
+        include.insert("b".into());
+        include.insert("c".into());
+
+        let order = topo_order(&map, &include).unwrap();
+        assert_eq!(order, vec!["a", "b", "c"]);
+    }
+}
